Update patch set 2
Patch Set 2:

If that is going to be used, it needs to be communicated clearly to Dominik for him and the iOS guys to adapt it. It will then also break existing customer-branded applications in the store, no? So honestly I would stick to the nick, or at allow it for backwards compatibility.

Patch-set: 2