Update patch set 4
Patch Set 4:

> > If we're not happy with this solution, another option (as
 > suggested by Guillem, thx!) would be to split the conflicting
 > modules into a separate package.
 > 
 > Honestly I am not entirely happy about this approach.
 > I am not 100% sure why we have conflict here. Because our
 > prosody.git repo ships the same modules new version prosody ships
 > in buster?

As far as I understand it, yes.

 > In this case, IMHO we should delete all the local modules in branch
 > master/buster and use it in buster+ release. Isn't it?

This means that we have to keep removing (or adding?!) modules when the package in Debian does so, which is a bit more work than the current approach.

What I don't understand very well is that, if the package name of our prosody is the same (is it?), why it's not taken as a new version of the prosody in Debian, and it's taken instead to be a different package.  AFAIK the "trying to overwrite" only happens when the package names are different.

Patch-set: 4