Update patch set 4
Patch Set 4: Code-Review+1

> What I don't understand very well is that, if the package name of
 > our prosody is the same (is it?), why it's not taken as a new
 > version of the prosody in Debian, and it's taken instead to be a
 > different package.  AFAIK the "trying to overwrite" only happens
 > when the package names are different.

Oh I see now, it's because of the different package name: ngcp-prosody-modules

We could use Break/Replaces, but not sure if it'll cause other complications and more harm than good in the long run.

So: I'm happy with the solution presented so far by mprokop, specially because it's not time consuming and unlikely to cause big problems or complications that cannot be solved just by adding/removing lines to that list.  But if people are not happy and want to explore other solutions, I am not opposed.

Patch-set: 4
Reviewer: Manuel Montecelo <1000042@85315e57-5f03-49c3-83f8-201ae2313a75>
Label: Code-Review=+1