TT#49205 Fix disco info vs items handling in cluster
Return proper disco#info when being asked for in a carrier environment,
instead of unconditionally returning disco#items in all disco requests.

To be done: also store room aliases in redis somewhere, as currently
we're losing this information along the way in carrier environments
when being asked for by a client.

Change-Id: If9497d0978808b7cfc94693a78566a74de5439f9