Revert "TT#44917 Ship certain modules only on Debian/stretch"
This reverts commit 039086e505aff4aa7ca8617edcecdccf4b7feef7.

The are modules not adapted for the new version of prosody shipping with buster
(0.11), so we have to keep using prosody 0.9 from stretch until this is sorted
out.

Thus reverting the commit, with a minor adaptation for the other changed that
creates a conflict (switching dependencies from debhelper>=10 to
debhelper-compat=12).

Change-Id: I6abd1e81dd432f4f1033c85490a232f0a0675412