MT#58471 Fix mod_carbons_adhoc call to adhoc module
The commit fix the startup error:

modulemanager: Error initializing module 'carbons_adhoc' on '...':
/usr/lib/prosody/modules/adhoc/adhoc.lib.lua:25: adhoc.new()
expects a permission argument, none given

The 'adhoc.new()' function now requires a new paramenter called
'permission' and this is not updated in the carbon plugin.
For the moment the premission has been statically set to
'any' value.

Change-Id: I897e421b994aae6ad95c29ab405a24c796d33e2b